Added hooks requirement for portals
