+++ src/renderer/hooks/use-toast-portal.ts
@@ -1,0 +1,26 @@
+import { useEffect, useState } from 'react';
+import { uuid } from 'renderer/utils';
+
+export const useToastPortal = () => {
+  const [loaded, setLoaded] = useState(false);
+  const [portalId] = useState(`toast-portal-${uuid()}`);
+
+  useEffect(() => {
+    const div = document.createElement('div');
+    div.id = portalId;
+    div.setAttribute(
+      'style',
+      'position: fixed; top: 10px; right: 10px; z-index: 9999'
+    );
+
+    document.getElementsByTagName('body')[0].prepend(div);
+
+    setLoaded(true);
+
+    return () => {
+      document.getElementsByTagName('body')[0].removeChild(div);
+    };
+  }, [portalId]);
+
+  return { loaded, portalId };
+};
